np2 copy False -> None

+++ src/aspire/utils/matrix.py
@@ -462,7 +462,7 @@ def nearest_rotations(A, allow_reflection=False):
         # If det(U)*det(V) = -1, we negate the third singular value to
         # ensure we have a rotation.
         neg_det_idx = np.linalg.det(U) * np.linalg.det(V) < 0
-        U[neg_det_idx] = U[neg_det_idx] @ np.diag((1, 1, -1)).astype(dtype, copy=False)
+        U[neg_det_idx] = U[neg_det_idx] @ np.diag((1, 1, -1)).astype(dtype, copy=None)
 
     rots = U @ V
